+++ crates/validator/src/common.rs
@@ -10,6 +10,7 @@ use e2etest_scylla_cluster::ScyllaClusterExt;
 use e2etest_scylla_cluster::ScyllaNodeConfig;
 use e2etest_scylla_proxy_cluster::ScyllaProxyClusterExt;
 use e2etest_scylla_proxy_cluster::ScyllaProxyNodeConfig;
+use e2etest_tls::TlsExt;
 use httpclient::HttpClient;
 use itertools::Itertools;
 use scylla::client::session::Session;
@@ -27,7 +28,6 @@ use std::time::Duration;
 use tap::Pipe;
 use tokio::time;
 use tracing::info;
-use vector_search_validator_tests::TlsExt;
 use vector_search_validator_tests::VectorStoreClusterExt;
 use vector_search_validator_tests::VectorStoreNodeConfig;
 use vector_store::IndexInfo;
