+++ crates/validator/src/lib.rs
@@ -27,6 +27,7 @@ use e2etest_firewall::Firewall;
 use e2etest_scylla_cluster::ScyllaCluster;
 use e2etest_scylla_cluster::ScyllaClusterExt;
 use e2etest_scylla_proxy_cluster::ScyllaProxyCluster;
+use e2etest_tls::Tls;
 use std::net::Ipv4Addr;
 use std::path::PathBuf;
 use std::sync::Arc;
@@ -38,7 +39,6 @@ use tracing_subscriber::filter;
 use tracing_subscriber::fmt;
 use tracing_subscriber::prelude::*;
 use vector_search_validator_tests::TestCase;
-use vector_search_validator_tests::Tls;
 use vector_search_validator_tests::VectorStoreCluster;
 use vector_search_validator_tests::VectorStoreClusterExt;
